+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-# ChangeLog
+# 0.1.1 - 12/08/2018
+### Added
++ Window title is now "Grigor Gargoyle"
 
-## 0.1.0 - 12/08/2018 - Initial Commit
+### Fixed
++ `goBack()` and `turn()` improved, now moving Grigor isn't as painful.
++ Entities can no longer exit the frame if `_can_exit_frame` is `False`
++ Hitbox now updates when the Entity moves
+
+# 0.1.0 - 12/07/2018 - Initial Commit
 ### Added
 + Entity, Player, Hitbox and Direction classes are fully functional (but not complete).